[chore] extract common test logic
